Switched TCSPolicy to bound mode, to ensure thread affinity
That is, this makes sure that when a specific thread communicates with
the enclave, another thread cannot hijack the session with the enclave
and try to mess with it in some way. This should prevent some exploits.

This commit also changes some global variables to thread_local to further
protect data leakage between threads.

Additionally, TCSNum is increased to 8. This is mostly to accomodate the
Go test runner, but also to enable queries to be performed on running nodes.