Fix IBC messages drop from the mempool
There's a mempool optimization in IBC where a tx doesn't enter the block if one of the messages failed in the mempool (yeah they execute the messages). On Secret if a message goes to CosmWasm, we verify the input against the signed bytes and against the signed block commit. The issue was that it couldn't find an IBC ack message in the signed block. Turned out to be a simulation in the mempool so there isn't a signed block at this point.

https://github.com/cosmos/ibc-go/blob/v4.3.1/modules/core/ante/ante.go#L35
https://github.com/cosmos/ibc-go/blob/v4.3.1/modules/core/ante/ante.go#L45
https://github.com/cosmos/ibc-go/blob/v4.3.1/modules/core/ante/ante.go#L55